// ui/src/app/enums/re-run-type.enum.ts
/*
 * *****************************************************************************
 *  Copyright (C) 2020 Testsigma Technologies Inc.
 *  All rights reserved.
 *  ****************************************************************************
 */

export enum ReRunType {
  NONE = "NONE",
  ALL_TESTS = "ALL_TESTS",
  ONLY_FAILED_TESTS = "ONLY_FAILED_TESTS"
}